phase782-build-fix-finale-closeout(hygiene+rhi+ui): cascade EXCLUDE_FROM_INSTALL + spirv multi-target + hb-ft.h + chord stub
Build was failing fresh-config after parallel FINALE-4/5/6/7/8 dispatch due to 5 cascading issues:

1. cd::asset::shader_cache / material_authoring / texture_compress + cd::render::scene_ingest
   needed EXCLUDE_FROM_INSTALL because their transitive PUBLIC deps (cd::shader, cd::material,
   bc7enc_rdo_lib, astcenc-none-static) are themselves EXCLUDE_FROM_INSTALL or FetchContent-only.

2. spirv_cross_glue used to assume vcpkg ships namespaced targets (spirv-cross-core::spirv-cross-core)
   but the current vcpkg actually exports plain targets (spirv-cross-core). Added a 4-way
   find_package() for core/glsl/hlsl/msl, plus a namespaced/plain target-detect fallback. The
   warning-suppression loop now skips IMPORTED targets (PRIVATE compile_options forbidden there).

3. HarfBuzz vcpkg promote (phase747) required hb-ft.h which is only built when the 'freetype'
   feature is enabled — flipped default-features:false → features:["freetype"].

4. HarfBuzzShaping.cpp had a namespace clash: file declared `namespace hb_shape { ... }` and
   then called the global `hb_shape(...)` function. Qualified the call as `::hb_shape(...)`.

5. apps/CMakeLists.txt unconditionally add_subdirectory(ios_demo) but the ios_demo/ directory
   was never committed (FINALE-7 W2A partial). Gated behind if(IOS AND EXISTS).

6. apps/editor/main.cpp (phase776 H3 chord) referenced ChordStateMachine + ChordModifier from
   the keyboard_shortcut_overlay panel but the panel never shipped those types. Added a stub
   close-out impl in the header (state machine + enum, full Sprint-2 deferred).
   The mods variable required an explicit static_cast<uint8_t>(ChordModifier::kCtrl).

Build clean. ctest 251/251 PASS. Working tree clean. Branch ready to push.

+++ apps/editor/main.cpp
@@ -3947,7 +3947,7 @@ int main(int argc, char** argv)
                         key_char = static_cast<char>('A' + (kc - kFirstAlpha));
                     }
                     const std::uint8_t mods =
-                        ctrl_down ? kso_ns::ChordModifier::kCtrl
+                        ctrl_down ? static_cast<std::uint8_t>(kso_ns::ChordModifier::kCtrl)
                                   : std::uint8_t{0U};
                     const std::string chord =
                         chord_sm.feed_key(mods, key_char, now_ms());
